move rcreate cmd tests to own module
